Switch Value numerics from BigDecimal to Double
Num, Dimen and Glue carry Double. The engine already computes layout
in Double, so BigDecimal bought arbitrary precision only between the
parser and the first getNumber call, at the cost of boxing and slower
arithmetic in every expression the macro processor evaluates.

+++ shared/src/main/scala-3/io/github/edadma/typesetter/Typesetter.scala
@@ -290,8 +290,8 @@ abstract class Typesetter:
     case v                     => sys.error(s"variable '$name' is not glue: ${Value.display(v)}")
 
   infix def getNumber(name: String): Double = getVar(name) match
-    case Value.Num(n)   => n.toDouble
-    case Value.Dimen(p) => p.toDouble
+    case Value.Num(n)   => n
+    case Value.Dimen(p) => p
     case v              => sys.error(s"variable '$name' is not a number: ${Value.display(v)}")
 
   def set(name: String, value: Any): Unit =
